requirements.txt
torch==1.5.1+cpu
opencv_contrib_python==4.2.0.34
Flask_SocketIO==4.3.1
numpy==1.19.0
Flask==1.1.2
facenet_pytorch==2.3.0
Pillow==7.2.0
gevent==20.6.2
gevent-websocket==0.10.1
python-engineio==3.13.1
python-socketio==4.6.0
